db: move excise to a separate file
Move DB.excise to its own file. Additionally, refactor the accumulation of EFOS
protected ranges of EFOSes that overlap the excise span by moving it into a
helper function in excise.go.
